🚀️ Completed Switch to Player Input Based Input
This will result in a reduction of time wasted as the C# calass is no longer required and the compiler will not need to recompile the project every time a binding or behaviour is changed.
This results higher ease of use as a GUI can now be used to tie buttons to fuctions.